+++ protocol/clproto_cpp/test/tests/test_parameter_proto.cpp
@@ -4,8 +4,32 @@
 #include <state_representation/space/cartesian/CartesianState.hpp>
 
 #include "clproto.h"
+#include "test_encode_decode.hpp"
 
 using namespace state_representation;
+
+template<typename T> using ParamT = std::tuple<T, clproto::ParameterMessageType>;
+
+static std::tuple<ParamT<bool>,
+                  ParamT<std::vector<bool>>,
+                  ParamT<int>,
+                  ParamT<std::vector<int>>,
+                  ParamT<double>,
+                  ParamT<std::vector<double>>,
+                  ParamT<std::string>,
+                  ParamT<std::vector<std::string>>,
+                  ParamT<Eigen::VectorXd>,
+                  ParamT<Eigen::MatrixXd>> parameter_test_cases{
+    std::make_tuple(false, clproto::ParameterMessageType::BOOL),
+    std::make_tuple(std::vector<bool>({true, false, true}), clproto::ParameterMessageType::BOOL_ARRAY),
+    std::make_tuple(1, clproto::ParameterMessageType::INT),
+    std::make_tuple(std::vector<int>({1, 2, 3}), clproto::ParameterMessageType::INT_ARRAY),
+    std::make_tuple(1.0, clproto::ParameterMessageType::DOUBLE), std::make_tuple(
+        std::vector<double>({1.0, 2.0, 3.0}), clproto::ParameterMessageType::DOUBLE_ARRAY
+    ), std::make_tuple("test", clproto::ParameterMessageType::STRING), std::make_tuple(
+        std::vector<std::string>({"1", "2", "3"}), clproto::ParameterMessageType::STRING_ARRAY
+    ), std::make_tuple(Eigen::VectorXd::Random(2), clproto::ParameterMessageType::VECTOR),
+    std::make_tuple(Eigen::MatrixXd::Random(2, 2), clproto::ParameterMessageType::MATRIX)};
 
 template<typename T>
 static void test_parameter_equal(const T& send_state, const T& recv_state) {
@@ -15,110 +39,30 @@ static void test_parameter_equal(const T& send_state, const T& recv_state) {
 }
 
 template<typename T>
-static void test_encode_decode_parameter(
-    const T& send_state, clproto::MessageType msg_type, clproto::ParameterMessageType param_type
-) {
-  std::string msg = clproto::encode(send_state);
-  EXPECT_TRUE(clproto::is_valid(msg));
-  EXPECT_EQ(clproto::check_message_type(msg), msg_type);
-  EXPECT_EQ(clproto::check_parameter_message_type(msg), param_type);
-
-  T recv_state("");
-  EXPECT_NO_THROW(clproto::decode<T>(msg));
-  EXPECT_TRUE(clproto::decode(msg, recv_state));
-  EXPECT_FALSE(recv_state.is_empty());
-
-  test_parameter_equal(send_state, recv_state);
-
-  auto send_state_ptr = make_shared_state(send_state);
-  msg = clproto::encode(send_state_ptr);
-  EXPECT_TRUE(clproto::is_valid(msg));
-  EXPECT_EQ(clproto::check_message_type(msg), msg_type);
-  EXPECT_EQ(clproto::check_parameter_message_type(msg), param_type);
-
-  T recv_state_2("");
-  auto recv_state_ptr = make_shared_state(recv_state_2);
-  EXPECT_NO_THROW(clproto::decode<std::shared_ptr<State>>(msg));
-  EXPECT_TRUE(clproto::decode(msg, recv_state_ptr));
-
-  recv_state_2 = *std::dynamic_pointer_cast<T>(recv_state_ptr);
-  test_parameter_equal(send_state, recv_state_2);
-}
-
-template<typename T>
-static void test_encode_decode_empty_parameter(const T& state) {
-  EXPECT_TRUE(state.is_empty());
-  std::string msg;
-  EXPECT_NO_THROW(msg = clproto::encode(state));
-
-  T recv_state("");
-  EXPECT_NO_THROW(recv_state = clproto::decode<T>(msg));
-  EXPECT_TRUE(recv_state.is_empty());
-}
+class ParameterProtoTest : public testing::Test {
+public:
+  ParameterProtoTest() : test_case_{std::get<ParamT<T>>(parameter_test_cases)} {}
+protected:
+  ParamT<T> test_case_;
+};
+TYPED_TEST_SUITE_P(ParameterProtoTest);
 
 TEST(ParameterProtoTest, EncodeParameterInterface) {
   auto send_state_ptr = make_shared_state(ParameterInterface("name", ParameterType::BOOL));
   EXPECT_THROW(std::string msg = clproto::encode(send_state_ptr), std::invalid_argument);
 }
 
-TEST(ParameterProtoTest, EncodeDecodeParameter) {
-  test_encode_decode_parameter(
-      Parameter<int>("A", 1), clproto::MessageType::PARAMETER_MESSAGE, clproto::ParameterMessageType::INT
-  );
-  test_encode_decode_parameter(
-      Parameter<std::vector<int>>("A", {1, 2, 3}), clproto::MessageType::PARAMETER_MESSAGE,
-      clproto::ParameterMessageType::INT_ARRAY
-  );
-  test_encode_decode_parameter(
-      Parameter<double>("A", 1.0), clproto::MessageType::PARAMETER_MESSAGE, clproto::ParameterMessageType::DOUBLE
-  );
-  test_encode_decode_parameter(
-      Parameter<std::vector<double>>("A", {1.0, 2.0, 3.0}), clproto::MessageType::PARAMETER_MESSAGE,
-      clproto::ParameterMessageType::DOUBLE_ARRAY
-  );
-  test_encode_decode_parameter(
-      Parameter<bool>("A", true), clproto::MessageType::PARAMETER_MESSAGE, clproto::ParameterMessageType::BOOL
-  );
-  test_encode_decode_parameter(
-      Parameter<std::vector<bool>>("A", {true, false, true, false}), clproto::MessageType::PARAMETER_MESSAGE,
-      clproto::ParameterMessageType::BOOL_ARRAY
-  );
-  test_encode_decode_parameter(
-      Parameter<std::string>("A", "value"), clproto::MessageType::PARAMETER_MESSAGE,
-      clproto::ParameterMessageType::STRING
-  );
-  test_encode_decode_parameter(
-      Parameter<std::vector<std::string>>("A", {"value 1", "value 2", "value 3"}),
-      clproto::MessageType::PARAMETER_MESSAGE, clproto::ParameterMessageType::STRING_ARRAY
-  );
-  Eigen::VectorXd vector_value(5);
-  vector_value << 1.0, 2.0, 3.0, 4.0, 5.0;
-  test_encode_decode_parameter(
-      Parameter<Eigen::VectorXd>("A", vector_value), clproto::MessageType::PARAMETER_MESSAGE,
-      clproto::ParameterMessageType::VECTOR
-  );
-  Eigen::MatrixXd matrix_value(2, 3);
-  matrix_value << 1.0, 2.0, 3.0, 4.0, 5.0, 6.0;
-  test_encode_decode_parameter(
-      Parameter<Eigen::MatrixXd>("A", matrix_value), clproto::MessageType::PARAMETER_MESSAGE,
-      clproto::ParameterMessageType::MATRIX
-  );
+TYPED_TEST_P(ParameterProtoTest, EncodeDecodeParameter) {
+  auto param = Parameter<TypeParam>("test", std::get<0>(this->test_case_));
+  clproto::test_encode_decode<Parameter<TypeParam>>(
+      param, clproto::PARAMETER_MESSAGE, test_parameter_equal<Parameter<TypeParam>>, std::get<1>(this->test_case_));
+  param.reset();
+  // FIXME(#50): Encoding of empty parameters is incorrect
+//  clproto::test_encode_decode<Parameter<TypeParam>>(
+//      param, clproto::PARAMETER_MESSAGE, test_parameter_equal<Parameter<TypeParam>>, std::get<1>(this->test_case_));
 }
 
-TEST(ParameterProtoTest, EncodeDecodeEmptyParameter) {
-  test_encode_decode_empty_parameter(Parameter<int>(""));
-  test_encode_decode_empty_parameter(Parameter<std::vector<int>>(""));
-  test_encode_decode_empty_parameter(Parameter<double>(""));
-  test_encode_decode_empty_parameter(Parameter<std::vector<double>>(""));
-  test_encode_decode_empty_parameter(Parameter<bool>(""));
-  test_encode_decode_empty_parameter(Parameter<std::vector<bool>>(""));
-  test_encode_decode_empty_parameter(Parameter<std::string>(""));
-  test_encode_decode_empty_parameter(Parameter<std::vector<std::string>>(""));
-  test_encode_decode_empty_parameter(Parameter<Eigen::VectorXd>(""));
-  test_encode_decode_empty_parameter(Parameter<Eigen::MatrixXd>(""));
-}
-
-TEST(MessageProtoTest, EncodeDecodeInvalidParameter) {
+TEST(ParameterProtoTest, EncodeDecodeInvalidParameter) {
   auto send_state = Parameter<CartesianState>("state", CartesianState::Random("state"));
   auto send_state_ptr = make_shared_state(send_state);
   EXPECT_THROW(clproto::encode(send_state_ptr), std::invalid_argument);
@@ -130,3 +74,17 @@ TEST(MessageProtoTest, EncodeDecodeInvalidParameter) {
   auto recv_state_ptr = make_shared_state(recv_state);
   EXPECT_FALSE(clproto::decode(msg, recv_state_ptr));
 }
+
+REGISTER_TYPED_TEST_SUITE_P(ParameterProtoTest, EncodeDecodeParameter);
+
+using ParameterTestTypes = testing::Types<bool,
+                                          std::vector<bool>,
+                                          int,
+                                          std::vector<int>,
+                                          double,
+                                          std::vector<double>,
+                                          std::string,
+                                          std::vector<std::string>,
+                                          Eigen::VectorXd,
+                                          Eigen::MatrixXd>;
+INSTANTIATE_TYPED_TEST_SUITE_P(Type, ParameterProtoTest, ParameterTestTypes);